Roles y permisos funcionando. Pendiente por agregar mas permisos con lo de Pablo.

+++ Proyecto/app/Http/Controllers/PermissionController.php
@@ -6,6 +6,7 @@ use Illuminate\Http\Request;
 use Spatie\Permission\Models\Permission;
 use App\Http\Requests\PermissionCreateRequest;
 use App\Http\Requests\PermissionEditRequest;
+use Illuminate\Support\Facades\Gate;
 
 class PermissionController extends Controller
 {
@@ -16,6 +17,7 @@ class PermissionController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('permission_index'), 403);
         $permissions = Permission::paginate(5);
         return view('permissions.index', compact('permissions'));
     }
@@ -27,6 +29,7 @@ class PermissionController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('permission_create'), 403);
         return view('permissions.create');
     }
 
@@ -61,6 +64,7 @@ class PermissionController extends Controller
      */
     public function edit(Permission $permission)
     {
+        abort_if(Gate::denies('permission_edit'), 403);
         return view('permissions.edit', compact('permission'));
     }
 
@@ -85,6 +89,7 @@ class PermissionController extends Controller
      */
     public function destroy(Permission $permission)
     {  
+        abort_if(Gate::denies('permission_destroy'), 403);
         $permission->delete();
         return back()->with('success', 'Permiso eliminado correctamente');
     }
