+++ Proyecto/app/Http/Controllers/RoleController.php
@@ -7,6 +7,7 @@ use Spatie\Permission\Models\Role;
 use Spatie\Permission\Models\Permission;
 use App\Http\Requests\RoleCreateRequest;
 use App\Http\Requests\RoleEditRequest;
+use Illuminate\Support\Facades\Gate;
 
 class RoleController extends Controller
 {
@@ -17,6 +18,7 @@ class RoleController extends Controller
      */
     public function index()
     {
+        abort_if(Gate::denies('role_index'), 403);
         $roles = Role::paginate(5);
         return view('roles.index', compact('roles'));
     }
@@ -28,6 +30,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        abort_if(Gate::denies('role_create'), 403);
         $permissions = Permission::all()->pluck('name', 'id');
         return view('roles.create', compact('permissions'));
     }
@@ -41,7 +44,7 @@ class RoleController extends Controller
     public function store(RoleCreateRequest $request)
     {
         $role = Role::create($request->only('name'));
-        $role->permissions()->sync($request->input('permissions', []));
+        $role->syncPermissions($request->input('permissions', []));
         return redirect()->route('roles.index')->with('success', 'Rol creado correctamente');
     }
 
@@ -64,6 +67,7 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        abort_if(Gate::denies('role_edit'), 403);
         $permissions = Permission::all()->pluck('name','id');
         $role->load('permissions');
         return view('roles.edit', compact('role', 'permissions'));
@@ -79,7 +83,7 @@ class RoleController extends Controller
     public function update(RoleEditRequest $request, Role $role)
     {
         $role->update($request->only('name'));
-        $role->permissions()->sync($request->input('permissions', []));
+        $role->syncPermissions($request->input('permissions', []));
         return redirect()->route('roles.index')->with('success', 'Rol actualizado correctamente');
     }
 
@@ -91,6 +95,7 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        abort_if(Gate::denies('role_destroy'), 403);
         $role->delete();
         return back()->with('success', 'Rol eliminado correctamente');
     }
